Restructure Substance Page to use type as a Prop
